+++ SPG.py
@@ -1,7 +1,4 @@
 # Simple Password Generator
 
 # Import SPM Python packages
-from SPM_CLI import *
-from SPM_GUI import *
-from SPM_API import *
 from SPM import *
