+++ SPM_API.py
@@ -1,7 +1,4 @@
 # Simple Password Manager Application Programming Interface
 
 # Import SPM Python packages
-from SPM_CLI import *
-from SPM_GUI import *
 from SPM import *
-from SPG import *
